Note that postgres requires a compatible locale to install
See http://dba.stackexchange.com/q/135127/91955

I was installing postgresql on Ubuntu 12.04 (Digital Ocean
droplet). The locale variables in my environment gets
transported to the DO box.

When postgres is installed via apt-get install postgresql, it
appears to succeed. However no default cluster is created, and
postgresql is not added as a service. This happens silently. I
realised what was going on when attempting to run
pg_createcluster manually failed.